requirements.txt
Django==2.2.8
djangorestframework==3.10.3
pytz==2019.3
sqlparse==0.3.1
dj-database-url==0.5.0
gunicorn==20.0.0
whitenoise==4.1.4
